+++ config.php
@@ -28,13 +28,15 @@ const NUVEI_PFW_PMS_REFUND_VOID    = array( 'cc_card', 'apmgw_expresscheckout' )
 
 const NUVEI_PFW_REST_ENDPOINT_INT   = 'https://ppp-test.nuvei.com/ppp/api/v1/';
 const NUVEI_PFW_REST_ENDPOINT_PROD  = 'https://secure.safecharge.com/ppp/api/v1/';
+const NUVEI_PFW_SDK_URL_PROD        = 'https://cdn.safecharge.com/safecharge_resources/v1/checkout/checkout.js';
+const NUVEI_PFW_SDK_URL_TAG         = 'https://devmobile.sccdev-qa.com/checkoutNext/checkout.js';
 const NUVEI_PFW_POPUP_AUTOCLOSE_URL = 'https://cdn.safecharge.com/safecharge_resources/v1/websdk/autoclose.html';
 
 const NUVEI_PFW_SESSION_OO_DETAILS      = 'nuvei_last_open_order_details'; // a session key
 const NUVEI_PFW_SESSION_PROD_DETAILS    = 'nuvei_order_details'; // products details
 const NUVEI_PFW_LOG_REQUEST_PARAMS      = 'Request params';
 
-define('NUVEI_PFW_SIMPLY_CONNECT_PATH', plugin_dir_url( __FILE__ ) . 'assets/js/nuveiSimplyConnect/');
+//define('NUVEI_PFW_SIMPLY_CONNECT_PATH', plugin_dir_url( __FILE__ ) . 'assets/js/nuveiSimplyConnect/');
 
 define(
 	'NUVEI_PFW_LOGS_DIR',